+++ apps/backend/src/libs/modules/mapbox/libs/enums/mapbox-api-path.enum.ts
@@ -1,0 +1,5 @@
+const MapboxAPIPath = {
+	DIRECTIONS: "/directions/v5",
+} as const;
+
+export { MapboxAPIPath };