+++ apps/backend/src/modules/routes/libs/enums/enums.ts
@@ -1,1 +1,2 @@
 export { RoutesExceptionMessage } from "./routes-exception-message.enum.js";
+export { RoutesApiPath } from "@smartscapes/shared";