+++ apps/backend/src/modules/routes/libs/types/type.ts
@@ -1,6 +1,8 @@
 export {
 	type RouteGetAllItemResponseDto,
 	type RouteGetByIdResponseDto,
+	type RoutesRequestConstructDto,
 	type RoutesRequestCreateDto,
 	type RoutesRequestPatchDto,
+	type RoutesResponseConstructDto,
 } from "@smartscapes/shared";