+++ apps/backend/src/modules/routes/routes.controller.ts
@@ -1,21 +1,23 @@
-import { RoutesApiPath } from "@smartscapes/shared";
-
-import { APIPath } from "~/libs/enums/enums.js";
+import { APIPath, HTTPCode } from "~/libs/enums/enums.js";
+import { setRateLimit } from "~/libs/hooks/hooks.js";
 import {
 	type APIHandlerOptions,
 	type APIHandlerResponse,
 	BaseController,
 } from "~/libs/modules/controller/controller.js";
-import { HTTPCode } from "~/libs/modules/http/http.js";
 import { type Logger } from "~/libs/modules/logger/logger.js";
 
+import { RoutesApiPath } from "./libs/enums/enums.js";
 import {
 	type RouteGetAllItemResponseDto,
 	type RouteGetByIdResponseDto,
+	type RoutesRequestConstructDto,
 	type RoutesRequestCreateDto,
 	type RoutesRequestPatchDto,
+	type RoutesResponseConstructDto,
 } from "./libs/types/type.js";
 import {
+	routesConstructValidationSchema,
 	routesCreateValidationSchema,
 	routesUpdateValidationSchema,
 } from "./libs/validation-schemas/validation-schemas.js";
@@ -25,6 +27,40 @@ import { type RoutesService } from "./routes.service.js";
  * @swagger
  * components:
  *   schemas:
+ *     Coordinate:
+ *       type: array
+ *       items:
+ *         type: number
+ *       minItems: 2
+ *       maxItems: 2
+ *       example: [30.5, 50.4]
+ *
+ *     GetMapboxRouteResponseDto:
+ *       type: object
+ *       properties:
+ *         routes:
+ *           type: array
+ *           items:
+ *             $ref: '#/components/schemas/MapboxRoute'
+ *         internalId:
+ *           type: string
+ *
+ *     MapboxRoute:
+ *       type: object
+ *       properties:
+ *         distance:
+ *           type: number
+ *         duration:
+ *           type: number
+ *         geometry:
+ *           type: object
+ *           properties:
+ *             coordinates:
+ *               type: array
+ *               items:
+ *                 $ref: '#/components/schemas/Coordinate'
+ *             type:
+ *               type: string
  *     Route:
  *       type: object
  *       properties:
@@ -47,7 +83,19 @@ class RoutesController extends BaseController {
 
 	public constructor(logger: Logger, routesService: RoutesService) {
 		super(logger, APIPath.ROUTES);
+
 		this.routesService = routesService;
+
+		const constructRequestsPerMinute = 5;
+		this.addRoute({
+			handler: this.constructRoute.bind(this),
+			method: "POST",
+			path: RoutesApiPath.CONSTRUCT,
+			preHandlers: [setRateLimit(constructRequestsPerMinute)],
+			validation: {
+				body: routesConstructValidationSchema,
+			},
+		});
 
 		this.addRoute({
 			handler: this.create.bind(this),
@@ -55,21 +103,25 @@ class RoutesController extends BaseController {
 			path: RoutesApiPath.ROOT,
 			validation: { body: routesCreateValidationSchema },
 		});
+
 		this.addRoute({
 			handler: this.delete.bind(this),
 			method: "DELETE",
 			path: RoutesApiPath.ID,
 		});
+
 		this.addRoute({
 			handler: this.find.bind(this),
 			method: "GET",
 			path: RoutesApiPath.ID,
 		});
+
 		this.addRoute({
 			handler: this.findAll.bind(this),
 			method: "GET",
 			path: RoutesApiPath.ROOT,
 		});
+
 		this.addRoute({
 			handler: this.patch.bind(this),
 			method: "PATCH",
@@ -80,10 +132,58 @@ class RoutesController extends BaseController {
 
 	/**
 	 * @swagger
+	 * /routes/construct:
+	 *   post:
+	 *     security:
+	 *       - bearerAuth: []
+	 *     tags:
+	 *       - Routes
+	 *     summary: Construct Mapbox route
+	 *     requestBody:
+	 *       required: true
+	 *       content:
+	 *         application/json:
+	 *           schema:
+	 *             type: object
+	 *             required:
+	 *               - pointsOfInterest
+	 *             properties:
+	 *               pointsOfInterest:
+	 *                 type: array
+	 *                 items:
+	 *                   type: integer
+	 *                 example: [1, 2]
+	 *     responses:
+	 *       200:
+	 *         description: Mapbox service response
+	 *         content:
+	 *           application/json:
+	 *             schema:
+	 *               type: object
+	 *               properties:
+	 *                 data:
+	 *                   $ref: '#/components/schemas/GetMapboxRouteResponseDto'
+	 */
+
+	public async constructRoute({
+		body: { pointsOfInterest },
+	}: APIHandlerOptions<{ body: RoutesRequestConstructDto }>): Promise<
+		APIHandlerResponse<RoutesResponseConstructDto>
+	> {
+		const data = await this.routesService.construct(pointsOfInterest);
+
+		return {
+			payload: { data },
+			status: HTTPCode.OK,
+		};
+	}
+
+	/**
+	 * @swagger
 	 * /routes:
 	 *   post:
 	 *     security:
-	 *      - bearerAuth: []
+	 *       - bearerAuth: []
 	 *     tags:
 	 *       - Routes
 	 *     summary: Create a new route
@@ -100,12 +200,15 @@ class RoutesController extends BaseController {
 	 *             properties:
 	 *               name:
 	 *                 type: string
+	 *                 example: Landscape alley
 	 *               description:
 	 *                 type: string
+	 *                 example: An alley with blooming flowers
 	 *               pois:
 	 *                 type: array
 	 *                 items:
 	 *                   type: number
+	 *                 example: [1, 2]
 	 *     responses:
 	 *       200:
 	 *         description: The created route
@@ -186,11 +289,15 @@ class RoutesController extends BaseController {
 	 *           type: string
 	 *     responses:
 	 *       200:
-	 *         description: Route was found succesfully
+	 *         description: Route was found successfully
 	 *         content:
 	 *           application/json:
 	 *             schema:
-	 *               $ref: '#/components/schemas/Route'
+	 *               type: object
+	 *               properties:
+	 *                 data:
+	 *                   type: object
+	 *                   $ref: '#/components/schemas/Route'
 	 */
 
 	public async find(
@@ -267,8 +374,10 @@ class RoutesController extends BaseController {
 	 *             properties:
 	 *               name:
 	 *                 type: string
+	 *                 example: Scenic walk
 	 *               description:
 	 *                 type: string
+	 *                 example: A calm stroll in countryside
 	 *     responses:
 	 *       200:
 	 *         description: Route updated successfully
@@ -278,16 +387,7 @@ class RoutesController extends BaseController {
 	 *               type: object
 	 *               properties:
 	 *                 data:
-	 *                   type: object
-	 *                   properties:
-	 *                     id:
-	 *                       type: number
-	 *                     name:
-	 *                       type: string
-	 *                       example: Landscape alley
-	 *                     description:
-	 *                       type: string
-	 *                       example: An alley with blooming flowers
+	 *                   $ref: '#/components/schemas/Route'
 	 */
 
 	public async patch(
@@ -310,5 +410,4 @@ class RoutesController extends BaseController {
 		};
 	}
 }
-
 export { RoutesController };