+++ apps/backend/src/modules/routes/routes.service.ts
@@ -2,9 +2,15 @@ import {
 	type RouteGetByIdResponseDto,
 	type RoutesRequestCreateDto,
 	type RoutesRequestPatchDto,
+	type RoutesResponseConstructDto,
 } from "@smartscapes/shared";
 
-import { HTTPCode } from "~/libs/modules/http/http.js";
+import { HTTPCode } from "~/libs/enums/enums.js";
+import {
+	MapboxAPIGeometric,
+	MapboxAPIProfile,
+	type MapboxDirectionsApi,
+} from "~/libs/modules/mapbox/mapbox.js";
 import { type CollectionResult, type Service } from "~/libs/types/types.js";
 
 import { type PointsOfInterestService } from "../points-of-interest/points-of-interest.service.js";
@@ -14,15 +20,43 @@ import { RoutesEntity } from "./routes.entity.js";
 import { type RoutesRepository } from "./routes.repository.js";
 
 class RoutesService implements Service {
+	private mapboxDirectionApi: MapboxDirectionsApi;
 	private pointsOfInterestService: PointsOfInterestService;
 	private routesRepository: RoutesRepository;
 
 	public constructor(
 		routesRepository: RoutesRepository,
 		pointsOfInterestService: PointsOfInterestService,
+		mapboxDirectionsApi: MapboxDirectionsApi,
 	) {
 		this.routesRepository = routesRepository;
 		this.pointsOfInterestService = pointsOfInterestService;
+		this.mapboxDirectionApi = mapboxDirectionsApi;
+	}
+
+	public async construct(
+		pointsOfInterest: number[],
+	): Promise<RoutesResponseConstructDto> {
+		const { items } = await this.pointsOfInterestService.findAll({
+			ids: pointsOfInterest,
+		});
+
+		if (items.length !== pointsOfInterest.length) {
+			throw new RoutesError({
+				message: RoutesExceptionMessage.POI_NOT_FOUND,
+				status: HTTPCode.NOT_FOUND,
+			});
+		}
+
+		const coordinates = items.map(({ location }) => location.coordinates);
+
+		const route = await this.mapboxDirectionApi.getRoute(
+			MapboxAPIProfile.WALKING,
+			coordinates,
+			MapboxAPIGeometric.GEOJSON,
+		);
+
+		return route;
 	}
 
 	public async create(