+++ packages/shared/src/modules/routes/libs/enums/routes-api-path.enum.ts
@@ -1,4 +1,5 @@
 const RoutesApiPath = {
+	CONSTRUCT: "/construct",
 	ID: "/:id",
 	ROOT: "/",
 } as const;