+++ packages/shared/src/modules/routes/libs/types/types.ts
@@ -1,5 +1,7 @@
+export { type RoutesRequestConstructDto } from "./routes-request-construct-dto.type.js";
 export { type RoutesRequestCreateDto } from "./routes-request-dto.type.js";
 export { type RoutesRequestPatchDto } from "./routes-request-update.dto.js";
+export { type RoutesResponseConstructDto } from "./routes-response-construct-dto.type.js";
 export {
 	type RouteGetAllItemResponseDto,
 	type RouteGetByIdResponseDto,