+++ packages/shared/src/modules/users/libs/types/user-auth-response-dto.type.ts
@@ -1,6 +1,10 @@
+import { type GroupItemWithPermissionsDto } from "../../../groups/types/types.js";
+
 type UserAuthResponseDto = {
 	email: string;
 	firstName: string;
+	group: GroupItemWithPermissionsDto;
+	groupId: number;
 	id: number;
 	lastName: string;
 };